Enable compiler warnings and fix most of them (#12)
This PR does not contain any new features. Still working on automating the LDPC-code quality (FER) simulations.
This PR fixes some minor bugs, as discovered by enabling compiler warnings.

* Bugfix: operator not-equals of RateAdaptiveCode was calling itself. Adds test for equality and inequality operators of RateAdaptiveCode.

* Enables compiler warnings. Fixes most of the warnings. Removes some obsolete code.

+++ benchmarks/code_simulation_helpers.hpp
@@ -35,7 +35,7 @@ namespace LDPC4QKD::CodeSimulationHelpers {
         for (auto i : in) {
             tmp += static_cast<double>(i);
         }
-        return tmp / in.size();
+        return tmp / static_cast<double>(in.size());
     }
